Refactor annotations (#59)
This PR refactors the the logic of the persistence layer handling the CRUD
operations of the asset annotations. This refactor it's needed as a previous
step to being able to implement CDC for the assets (including their
annotations). The concrete modifications are:

* Now all the CRUD operations over assets (or asset annotations) end up calling
the same methods, for creating, modifying or deleting assets.
* The logic for checking if the annotations of an asset belong to the
authorized team of the user calling the endpoints of the asset annotations, are
improved in order to avoid a race condition that could make the check to fail.
* Some test for the asset annotations storage layer have been added.
* Some comments about some possible concurrency problems when modifying assets
were added, so we can address them in future PRs.